# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AYA Registry API",
    "description": "Public API for the AYA Registry — an index of organizations rated for AI readability (AIO score 0-100). No authentication required. Rate limit: 30 requests/minute per IP.",
    "version": "1.0.0",
    "contact": {
      "name": "AI Visionary",
      "email": "hello@ai-visionary.xyz",
      "url": "https://ai-visionary.xyz"
    },
    "license": {
      "name": "CC-BY-4.0",
      "url": "https://creativecommons.org/licenses/by/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://ai-visionary.xyz/api/aya",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "searchEntities",
        "summary": "Search entities",
        "description": "Search by name, domain, sector or country. Sorted by relevance + certified bonus. Multi-word matching with stop-word filtering.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search term (company name, domain, sector, country code)",
            "schema": { "type": "string", "minLength": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Max results (1-200, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "q": { "type": "string" },
                    "n": { "type": "integer" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "country": { "type": "string" },
                          "sector": { "type": "string" },
                          "score": { "type": "integer" },
                          "certified": { "type": "boolean" },
                          "url": { "type": "string" },
                          "entity_id": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}